requirements.txt
fastapi
python-dotenv==0.15.0
psycopg2
geopy
uvicorn
requests
